refactor: move nvim-cmp sources from root
